ispravljene metode

+++ studsluzba_server/src/main/java/org/raflab/studsluzba/controllers/NastavnikZvanjeController.java
@@ -4,24 +4,35 @@ import lombok.RequiredArgsConstructor;
 import org.raflab.studsluzba.controllers.request.NastavnikZvanjeRequest;
 import org.raflab.studsluzba.controllers.response.NastavnikZvanjeResponse;
 import org.raflab.studsluzba.model.NastavnikZvanje;
+import org.raflab.studsluzba.repositories.NastavnikZvanjeRepository;
 import org.raflab.studsluzba.services.NastavnikZvanjeService;
 import org.raflab.studsluzba.utils.Converters;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/nastavnik-zvanje")
 @RequiredArgsConstructor
 public class NastavnikZvanjeController {
     private final NastavnikZvanjeService service;
-
+    private final NastavnikZvanjeRepository repository;
     @PostMapping("/add")
-    public Long add(@RequestBody @Valid NastavnikZvanjeRequest request) {
+    public ResponseEntity<?> add(@RequestBody @Valid NastavnikZvanjeRequest request) {
         NastavnikZvanje nz = service.save(request);
-        return nz.getId();
+
+        if (nz == null) {
+            return ResponseEntity
+                    .badRequest()
+                    .body("Greška: nastavnik ne postoji ili već ima aktivno zvanje koje ste uneli.");
+        }
+
+        return ResponseEntity.ok(nz.getId());
     }
+
 
     @GetMapping("/all")
     public List<NastavnikZvanjeResponse> getAll() {
@@ -29,13 +40,23 @@ public class NastavnikZvanjeController {
     }
 
     @GetMapping("/{id}")
-    public NastavnikZvanjeResponse getById(@PathVariable Long id) {
-        return Converters.toNastavnikZvanjeResponse(service.findById(id));
+    public ResponseEntity<?> getById(@PathVariable Long id) {
+        Optional<NastavnikZvanje> nz = repository.findById(id);
+
+        if (nz.isEmpty()) {
+            return ResponseEntity.badRequest().body("Zvanje sa ID " + id + " ne postoji.");
+        }
+
+        return ResponseEntity.ok(Converters.toNastavnikZvanjeResponse(nz.get()));
     }
 
     @DeleteMapping("/delete/{id}")
-    public String delete(@PathVariable Long id) {
+    public ResponseEntity<?> delete(@PathVariable Long id) {
+        if (!repository.existsById(id)) {
+            return ResponseEntity.badRequest().body("Ne postoji zvanje sa ID " + id);
+        }
         service.deleteById(id);
-        return "Obrisano zvanje sa ID " + id;
+        return ResponseEntity.ok("Obrisano zvanje sa ID " + id);
     }
+
 }
